requirements.txt
Django==4.0.7
django-crispy-forms==1.14.0
django-filter==2.2.0
djangorestframework==3.13.1
Markdown==3.4.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
django-extensions==3.2.0
django-taggit==3.0.0


# For using MySQL, find instructions at
#   https://pypi.org/project/mysqlclient/
mysqlclient==2.1.1